+++ embedded-hal-bus/src/spi/critical_section.rs
@@ -5,6 +5,7 @@ use embedded_hal::digital::OutputPin;
 use embedded_hal::spi::{ErrorType, Operation, SpiBus, SpiDevice};
 
 use super::DeviceError;
+use crate::spi::shared::transaction;
 
 /// `critical-section`-based shared bus [`SpiDevice`] implementation.
 ///
@@ -66,29 +67,7 @@ where
         critical_section::with(|cs| {
             let bus = &mut *self.bus.borrow_ref_mut(cs);
 
-            self.cs.set_low().map_err(DeviceError::Cs)?;
-
-            let op_res = operations.iter_mut().try_for_each(|op| match op {
-                Operation::Read(buf) => bus.read(buf),
-                Operation::Write(buf) => bus.write(buf),
-                Operation::Transfer(read, write) => bus.transfer(read, write),
-                Operation::TransferInPlace(buf) => bus.transfer_in_place(buf),
-                Operation::DelayUs(us) => {
-                    bus.flush()?;
-                    self.delay.delay_us(*us);
-                    Ok(())
-                }
-            });
-
-            // On failure, it's important to still flush and deassert CS.
-            let flush_res = bus.flush();
-            let cs_res = self.cs.set_high();
-
-            op_res.map_err(DeviceError::Spi)?;
-            flush_res.map_err(DeviceError::Spi)?;
-            cs_res.map_err(DeviceError::Cs)?;
-
-            Ok(())
+            transaction(operations, bus, &mut self.delay, &mut self.cs)
         })
     }
 }
